+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG for Sulu
 ==================
 
 * dev-develop
+    * FEATURE     #3078 [MediaBundle]         Added sort-by title for media-dataprovider
     * FEATURE     #3069 [PreviewBundle]       Added cache configuration for preview & websocket context
     * ENHANCEMENT #3071 [All]                 Updated willdurand/hateoas-bundle
     * FEATURE     #3066 [AutomationBundle]    Added notification-badge to automation-tab